+++ domoticz_button/src/sota.cpp
@@ -17,15 +17,36 @@ WiFiClient wifiClient;
 //#define USE_MAC
 
 #ifdef USE_MAC
-
 String getMAC() {
   uint8_t mac[6];
   char result[14];
   snprintf( result, sizeof( result ), "%02x%02x%02x%02x%02x%02x", mac[ 0 ], mac[ 1 ], mac[ 2 ], mac[ 3 ], mac[ 4 ], mac[ 5 ] );
   return String( result );
 }
-
 #endif
+
+bool otaUpdate(void) {
+  #ifdef USE_MAC
+  String mac = getMAC();
+  sendToLogPf(LOG_INFO, PSTR("Checking for firmware updates for %s"),  mac.c_str());
+  #else
+  String mac = config.hostname;
+  #endif
+  String url = String(config.otaUrlBase) + mac + ".binx"; // make sure it does not work
+  t_httpUpdate_return ret = ESPhttpUpdate.update(wifiClient, config.otaHost, config.otaPort, url);
+  switch (ret) {
+    case HTTP_UPDATE_FAILED:
+      sendToLogPf(LOG_ERR, PSTR("Firmware update failed with error (%d): %s"), ESPhttpUpdate.getLastError(), ESPhttpUpdate.getLastErrorString().c_str());
+      break;
+    case HTTP_UPDATE_NO_UPDATES:
+      sendToLogP(LOG_ERR, PSTR("Firmware update failed with error : NO UPDATES"));
+      break;
+    case HTTP_UPDATE_OK:
+      sendToLogPf(LOG_INFO, PSTR("Firware update succeeded"));
+      break;
+  }
+  return ret == HTTP_UPDATE_OK;
+}
 
 OTA_result_t checkForUpdates(void) {
   OTA_result_t result = OTA_FAILED;
@@ -35,12 +56,9 @@ OTA_result_t checkForUpdates(void) {
   #else
   String mac = config.hostname;
   #endif
-  String imageURL = String("http://") + config.otaHost + ':' + config.otaPort + config.otaUrlBase + mac; 
-  String versionURL = imageURL + ".version";
-  imageURL.concat(".bin");
+  String versionURL = String("http://") + config.otaHost + ':' + config.otaPort + config.otaUrlBase + mac + '.version'; 
   
   HTTPClient httpClient;
-
   httpClient.begin(wifiClient, versionURL);
   int httpCode = httpClient.GET();
   if (httpCode == 404) {
@@ -50,30 +68,16 @@ OTA_result_t checkForUpdates(void) {
   } else { 
     String newFWVersion = httpClient.getString();
     int newVersion = newFWVersion.toInt();
-    //sendToLogPf(LOG_INFO, PSTR("Current firmware version %d, available version %d"), VERSION, newVersion);
-
     if (newVersion <= VERSION) {
       sendToLogPf(LOG_INFO, PSTR("Current firmware version %d is the latest version"), VERSION);
       result = OTA_NO_NEW_VERSION;
-    } else {  
-      HTTPUpdateResult ret = ESPhttpUpdate.update(imageURL);
-      switch (ret) {
-          case HTTP_UPDATE_FAILED:
-            sendToLogPf(LOG_ERR, PSTR("Firmware update to version %d failed with error (%d): %s"), newVersion, ESPhttpUpdate.getLastError(), ESPhttpUpdate.getLastErrorString().c_str());
-            break;
-
-          case HTTP_UPDATE_NO_UPDATES:
-            sendToLogP(LOG_ERR, PSTR("Firmware update to version %d failed: NO UPDATES"));
-            break;
-
-          case HTTP_UPDATE_OK:
-            sendToLogPf(LOG_INFO, PSTR("Updated firmware from version %d to version %d"), VERSION, newVersion);
-            result = OTA_NEW_VERSION_LOADED;
-            break;
-      }
-    }
+    } else if (otaUpdate()) 
+      result = OTA_NEW_VERSION_LOADED;
   }
   httpClient.end();
   return result;
 }
+
+
+