+++ domoticz_button/src/sota.h
@@ -1,6 +1,8 @@
 #ifndef SOTA_H
 #define SOTA_H
 
+#include <Arduino.h>
+
 // Self updating over the air
 
 // Types of Domoticz virtual devices that will be handled 
@@ -11,5 +13,7 @@ enum OTA_result_t {
 };
 
 OTA_result_t checkForUpdates(void);
+
+bool otaUpdate(void);
 
 #endif